Cleaner interface
