+++ test/test.html
@@ -1,18 +1,14 @@
-<html>
-  <body>
-    <script src="../xmllint.js"></script>
-    <form onsubmit="console.log(xmllint.validateXML({ xml: xml.value, schema: schema.value}).errors); return false;">
-      XML: <textarea rows=10 cols=80 id="xml"></textarea>
-      <br>
-      Schema: <textarea rows=10 cols=80 id="schema"></textarea>
-      <br>
-      <input type="submit" value="validate">
-    </form>
-    <!-- browserify/browserify-shim bundle test:
-      expect to see:
-        Object { errors: null } 
-      in the console log.
-    -->
-    <script src="./bundle.js"></script>
-  </body>
-</html>
+<!doctype html>
+<meta charset="utf-8">
+<title>xmllint dtd test</title>
+<link rel="stylesheet" href="test.css">
+
+<body>
+
+<div id="status">Loading…</div>
+
+<pre id="lint-container"><code id="lint"></code></pre>
+<pre id="xml-container"><code id="xml"></code></pre>
+
+<script src="../xmllint.js"></script>
+<script src="test.js"></script>
