+++ test/test.xsd
@@ -63,4 +63,3 @@
  </xsd:simpleType>
 
 </xsd:schema>
-
